[Refactoring] [Move Only] Move wallet files into their own folder
