+++ src/qt/pivx.cpp
@@ -34,7 +34,7 @@
 #include "util.h"
 
 #ifdef ENABLE_WALLET
-#include "wallet.h"
+#include "wallet/wallet.h"
 #endif
 
 #include <stdint.h>
