+++ src/qt/transactiondesc.cpp
@@ -18,7 +18,7 @@
 #include "timedata.h"
 #include "ui_interface.h"
 #include "util.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <stdint.h>
 #include <string>
