+++ src/test/rpc_wallet_tests.cpp
@@ -7,7 +7,7 @@
 #include "rpc/client.h"
 
 #include "base58.h"
-#include "wallet.h"
+#include "wallet/wallet.h"
 
 #include <boost/algorithm/string.hpp>
 #include <boost/test/unit_test.hpp>
