+++ src/test/zerocoin_transactions_tests.cpp
@@ -7,8 +7,8 @@
 #include "chainparams.h"
 #include "coincontrol.h"
 #include "main.h"
-#include "wallet.h"
-#include "walletdb.h"
+#include "wallet/wallet.h"
+#include "wallet/walletdb.h"
 #include "txdb.h"
 #include <boost/test/unit_test.hpp>
 #include <iostream>
